+++ src/runtime/ui.config/overlays/notification.ts
@@ -6,6 +6,7 @@ export default {
   inner: 'w-0 flex-1',
   title: 'text-sm font-medium text-gray-900 dark:text-white',
   description: 'mt-1 text-sm leading-4 text-gray-500 dark:text-gray-400',
+  descriptionOnly: 'mt-0 leading-5',
   actions: 'flex items-center gap-2 mt-3 flex-shrink-0',
   background: 'bg-white dark:bg-gray-900',
   shadow: 'shadow-lg',